requirements.txt
groq
gradio
python-dotenv
langchain
langchain-community
langchain_text_splitters
langchain-openai
langchain-huggingface
langchain-groq
langchain-chroma
langchain-core
chromadb
pypdf
pymupdf
scikit-learn
sentence-transformers
ticker
numpy
plotly